[Nodes] Make lengths input of `BatchOneHot` int32
Description:
This commit modifies the `BatchOneHot` operator to use 32-bit integer
values in the `lengths` input rather than 64-bit integers. This change
was supposed to be part of an earlier commit that migrated the index
inputs of all sparse operators to 32-bit integer, but it was overlooked
at the time.

Testing:
This commit modifies the operator test for `BatchOneHot` to use 32-bit
integers. The modified test passes, as do all other unit tests.

+++ lib/Backends/Interpreter/InterpreterNodes.cpp
@@ -898,7 +898,7 @@ void InterpreterFunction::fwdScatterAssignInst(
 template <typename ElemTy>
 void InterpreterFunction::fwdBatchOneHotImpl(const glow::BatchOneHotInst *I) {
   auto dataH = getWeightHandle<ElemTy>(I->getData());
-  auto lengthsH = getWeightHandle<int64_t>(I->getLengths());
+  auto lengthsH = getWeightHandle<int32_t>(I->getLengths());
   auto valuesH = getWeightHandle<ElemTy>(I->getValues());
   auto destH = getWeightHandle<ElemTy>(I->getDest());
